Redistribute functions within dbUpdater, processRepo, and dataAggregation.ts
