+++ packages/graphql_server/src/dataAggregation.ts
@@ -1,8 +1,14 @@
 import { GitHubOrganization, GitHubInfo, GitHubUser } from '../types/githubApi'
 import { getOrganizationInfo, getRepoInfo, getUserInfo } from './api/githubApi'
-import { OrganizationInsertion, PersonInsertion } from '../types/dataAggregation'
+import {
+  OrganizationInsertion,
+  OrganizationUpdate,
+  PersonInsertion,
+  ProjectUpdate
+} from '../types/dataAggregation'
 import supabase from './supabase'
 import { StarRecord } from '../types/starHistory'
+import { LinkedInCompanyProfile } from '../types/linkedInScraping'
 
 /**
  * Formats the github data into a format that can be inserted into the database.
@@ -31,6 +37,122 @@ export const turnIntoProjectInsertion = (githubData: GitHubInfo, starHistory: St
     star_history: starHistory,
     is_bookmarked: false
   }
+}
+
+/**
+ * Formats the linkedInData to the format that is used in the db
+ * @param {LinkedInCompanyProfile} linkedInData - The companny linkedIn data
+ */
+export const formatLinkedInCompanyData = (
+  linkedInData: LinkedInCompanyProfile
+): OrganizationUpdate => {
+  return {
+    crunchbase: linkedInData.crunchbaseUrl,
+    founded: parseInt(linkedInData.founded, 10),
+    hq_location: linkedInData.hqLocation,
+    industries: linkedInData.industries,
+    linkedin_about: linkedInData.about,
+    linkedin_followers: linkedInData.followers,
+    linkedin_updates: linkedInData.updates,
+    linkedin_url: linkedInData.url,
+    linkedin_website_url: linkedInData.website,
+    number_of_employees: parseInt(linkedInData.employeesAmountInLinkedin),
+    specialties: linkedInData.specialties
+  }
+}
+
+/**
+ * Updates the supabase entry of a repo
+ * @param {string} name - The name of the repo.
+ * @param {string} owner - The name of the owner of the repo.
+ * @param {ProjectUpdate} updatedProject - The Changes that should be put on supabase
+ * @returns {boolean} - Whether the update was successful
+ */
+export const updateSupabaseProject = async (
+  name: string,
+  owner: string,
+  updatedProject: ProjectUpdate
+) => {
+  //check whether the repo is in the db
+  if (!(await repoIsAlreadyInDB(name, owner))) {
+    return false
+  }
+  const owningOrganizationID = await getOrganizationID(owner)
+
+  const { error: ownerUpdateError } = await supabase
+    .from('project')
+    .update(updatedProject)
+    .eq('name', name)
+    .eq('owning_organization', owningOrganizationID)
+
+  if (!ownerUpdateError) return true
+  const owningPersonID = await getPersonID(owner)
+  const { error: ownerUpdateError2 } = await supabase
+    .from('project')
+    .update(updatedProject)
+    .eq('name', name)
+    .eq('owning_person', owningPersonID)
+
+  return ownerUpdateError2 ? false : true
+}
+
+/**
+ * Sets the trending states of all projects to false.
+ * contributor_count comparison because I want to target all rows but have to specify a filter
+ */
+export const purgeTrendingState = async () => {
+  const { error: supabaseError } = await supabase
+    .from('project')
+    .update({ is_trending_daily: false, is_trending_weekly: false, is_trending_monthly: false })
+    .neq('contributor_count', -1000)
+  supabaseError && console.error('Error while purging trending state: \n', supabaseError)
+}
+
+/**
+ * Checks if the repo is already in the db.
+ * @param {string} name - The name  of the repo.
+ * @param {string} owner - The name of the owner of the repo.
+ * @returns {boolean} True if the repo is already in the db.
+ */
+export const repoIsAlreadyInDB = async (name: string, owner: string) => {
+  // check if there are repositories with the same name
+  const { data: matchingRepos, error: checkRepoIfRepoInDBError } = await supabase
+    .from('project')
+    .select('*')
+    .eq('name', name)
+  checkRepoIfRepoInDBError &&
+    console.error(
+      'Error while checking if',
+      name,
+      'is in the database: \n',
+      checkRepoIfRepoInDBError
+    )
+  // if there are no repositories with the same name return false
+  if (!matchingRepos) return false
+
+  // for each of those with the same name check if the owner has the same name
+  for (const repo of matchingRepos) {
+    // if the owner is an organization
+    if (repo.owning_organization) {
+      const { data: owning_organization } = await supabase
+        .from('organization')
+        .select('*')
+        .eq('id', repo.owning_organization)
+
+      // the owner has the same name -> the repo is already in the database
+      if (owning_organization?.[0]?.login === owner) return true
+    } else {
+      const { data: owning_person } = await supabase
+        .from('associated_person')
+        .select('*')
+        .eq('id', repo.owning_person)
+
+      // the owner has the same name -> the repo is already in the database
+      if (owning_person?.[0]?.login === owner) return true
+    }
+  }
+
+  return false
 }
 
 /**
